Multiple gui support
It seems to me that these changes allow the WrenchFu to be used by several modules at the same time and not have unnessessary clashes. 
Before, the mod would close all other guis opened with it. yet since there are no specifications on what is drawn by caller mod, this may be unwise.
Now mods should explicitly provide the list of entities handlers they wish not to be openned together. 
Also the max_distance  setting was changed to be mod specific